split mocked response console tests
